Refactor code into modular components
