+++ GPS Logger/LocationManager.swift
@@ -1,0 +1,153 @@
+import Foundation
+import CoreLocation
+import UIKit
+import Combine
+
+/// Handles location updates and recording of log entries and photos.
+final class LocationManager: NSObject, ObservableObject, CLLocationManagerDelegate {
+    private let locationManager = CLLocationManager()
+
+    let flightLogManager: FlightLogManager
+    private let altitudeFusionManager: AltitudeFusionManager
+    let settings: Settings
+
+    @Published var lastLocation: CLLocation?
+    @Published var isRecording = false
+
+    @Published var rawGpsAltitude: Double = 0.0
+    @Published var rawGpsAltitudeChangeRate: Double = 0.0
+    private var previousRawAltitudeTimestamp: Date?
+
+    @Published var declination: Double = 0.0
+
+    var photoCounter: Int = 0
+    var pendingPhotoIndex: Int? = nil
+    var logTimer: Timer?
+
+    private var cancellables = Set<AnyCancellable>()
+
+    init(flightLogManager: FlightLogManager,
+         altitudeFusionManager: AltitudeFusionManager,
+         settings: Settings) {
+        self.flightLogManager = flightLogManager
+        self.altitudeFusionManager = altitudeFusionManager
+        self.settings = settings
+        super.init()
+
+        UIDevice.current.isBatteryMonitoringEnabled = true
+        locationManager.delegate = self
+        locationManager.desiredAccuracy = kCLLocationAccuracyBestForNavigation
+        locationManager.distanceFilter = kCLDistanceFilterNone
+        locationManager.requestAlwaysAuthorization()
+
+        if CLLocationManager.headingAvailable() {
+            locationManager.startUpdatingHeading()
+        }
+
+        NotificationCenter.default.addObserver(self, selector: #selector(didEnterBackground), name: UIApplication.didEnterBackgroundNotification, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(willEnterForeground), name: UIApplication.willEnterForegroundNotification, object: nil)
+
+        settings.$logInterval
+            .sink { [weak self] newInterval in
+                guard let self else { return }
+                if self.isRecording {
+                    self.logTimer?.invalidate()
+                    self.logTimer = Timer.scheduledTimer(withTimeInterval: newInterval, repeats: true) { [weak self] _ in
+                        self?.recordLog()
+                    }
+                }
+            }
+            .store(in: &cancellables)
+    }
+
+    deinit {
+        NotificationCenter.default.removeObserver(self)
+    }
+
+    @objc private func didEnterBackground() {
+        if !isRecording {
+            locationManager.stopUpdatingLocation()
+        }
+    }
+
+    @objc private func willEnterForeground() {
+        startUpdatingForDisplay()
+    }
+
+    func startUpdatingForDisplay() {
+        locationManager.startUpdatingLocation()
+        locationManager.allowsBackgroundLocationUpdates = true
+        locationManager.pausesLocationUpdatesAutomatically = false
+    }
+
+    func startRecording() {
+        flightLogManager.startSession()
+        isRecording = true
+
+        rawGpsAltitude = 0.0
+        rawGpsAltitudeChangeRate = 0.0
+        previousRawAltitudeTimestamp = nil
+
+        logTimer = Timer.scheduledTimer(withTimeInterval: settings.logInterval, repeats: true) { [weak self] _ in
+            self?.recordLog()
+        }
+    }
+
+    func stopRecording() {
+        isRecording = false
+        logTimer?.invalidate()
+        logTimer = nil
+        altitudeFusionManager.stopUpdates()
+        flightLogManager.endSession()
+    }
+
+    func recordPhotoCapture() -> Int? {
+        if !isRecording { return nil }
+        photoCounter += 1
+        pendingPhotoIndex = photoCounter
+        return photoCounter
+    }
+
+    func recordLog() {
+        guard let loc = lastLocation else { return }
+        let altitudeFt = loc.altitude * 3.28084
+        let now = Date()
+        var vspeed = rawGpsAltitudeChangeRate
+        if let prevTimestamp = previousRawAltitudeTimestamp {
+            let dt = now.timeIntervalSince(prevTimestamp)
+            if dt > 0 {
+                let change = altitudeFt - rawGpsAltitude
+                vspeed = change / dt * 60
+            }
+        }
+        previousRawAltitudeTimestamp = now
+        rawGpsAltitude = altitudeFt
+        rawGpsAltitudeChangeRate = vspeed
+        altitudeFusionManager.latestGpsAltitude = altitudeFt
+        altitudeFusionManager.rawGpsVerticalSpeed = vspeed
+        altitudeFusionManager.gpsVerticalAccuracy = loc.verticalAccuracy * 3.28084
+        altitudeFusionManager.startUpdates(gpsAltitude: altitudeFt)
+
+        let log = FlightLog(timestamp: now,
+                             latitude: loc.coordinate.latitude,
+                             longitude: loc.coordinate.longitude,
+                             gpsAltitude: altitudeFt,
+                             speedKt: loc.speed * 1.94384,
+                             magneticCourse: loc.course,
+                             horizontalAccuracyM: loc.horizontalAccuracy,
+                             verticalAccuracyFt: loc.verticalAccuracy * 3.28084,
+                             altimeterPressure: nil,
+                             rawGpsAltitudeChangeRate: rawGpsAltitudeChangeRate,
+                             relativeAltitude: altitudeFusionManager.relativeAltitude ?? 0,
+                             barometricAltitude: altitudeFusionManager.baselineAltitude.map { $0 + (altitudeFusionManager.relativeAltitude ?? 0) } ?? altitudeFt,
+                             latestAcceleration: altitudeFusionManager.latestAcceleration,
+                             fusedAltitude: altitudeFusionManager.fusedAltitude ?? altitudeFt,
+                             fusedAltitudeChangeRate: altitudeFusionManager.altitudeChangeRate,
+                             baselineAltitude: altitudeFusionManager.baselineAltitude,
+                             measuredAltitude: altitudeFusionManager.measuredAltitude,
+                             kalmanUpdateInterval: altitudeFusionManager.kalmanUpdateInterval,
+                             photoIndex: pendingPhotoIndex)
+        pendingPhotoIndex = nil
+        flightLogManager.addLog(log)
+    }
+}
